Fixing nits from code review
