+++ src/plugins/telemetry_management_section/tsconfig.json
@@ -5,6 +5,7 @@
     "isolatedModules": true
   },
   "include": [
+    "common/**/*",
     "public/**/*",
     "../../../typings/**/*"
   ],
